Fix use defaultButton from props when toastClose is true
see https://github.com/fkhadra/react-toastify/pull/313#issuecomment-513437356

that change had made it a lot harder to use the default button from the props. This change is there to make it easy to create toast using the button defined in the container. If people want the default button they can easily import it from the callspot.