Fix clear all bug when promise has neither a pending nor a success/error
On resolving a promise, if there is no success/error, then the pending toast needs to be dismissed. But if there was no pending toast, then toast.dismiss() is being called with undefined, which clears all toast. It should only call toast.dimiss() if id exists for the pending toast.